Groups helpers with domains for `TecoServiceGenerator`
