+++ Sources/TecoServiceGenerator/helpers/ActionHelpers.swift
@@ -1,0 +1,4 @@
+func skipAuthorizationParameter(for action: String) -> String {
+    // Special rule for sts:AssumeRoleWithSAML & sts:AssumeRoleWithWebIdentity
+    return action.hasPrefix("AssumeRoleWith") ? ", skipAuthorization: true" : ""
+}
